+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - app/Support/Collection
 - ProductController getProductsByCategory method
+- vines-with-products api route
+- ApiControllerService getWithProducts, paginate methods
+- VineController getWithProducts method
 
 ## [0.0.2] - 2020-08-05
 ### Added